[monad-event-ring] add event-ring Rust version
monad execution is gaining an event notification system. It can
communicate with external processes via a shared memory communication
mechanism, described in the documentation in the `monad` repository.
This will initially be used by the RPC server to support WebSocket
realtime events.

This commit adds support for Rust clients. It reuses part of the
C/C++ client library via FFI, but the performance sensitive part
(event_reader.rs) is a native Rust reimplementation.