[monad-exec-events] add execution ring events
This Rust package defines the execution events. An execution event is a
notification that the Ethereum Virtual Machine has performed some action,
such as emitting a transaction log or changing an account balance. These
events are produced by the execution daemon, a separate process running on
the same host. These events are consumed using the `monad-event-ring`
package.