Start vote pacing timer on startup
The vote pacing timer would previously only start after a proposal
carrying a TC was received. This changes the behavior to immediately
fire the vote pacing timer on startup.

This is theoretically slightly inaccurate, as there's no guarantee that
the vote pacing interval elapsed since the last time the node voted, but
in practice this will not happen.